Revert "Merge pull request #1369 from jsantell/961194-modularize-loader-utils"
This reverts commit aaf3c9061f1b8cc97deca7417937f029b9d1ea75, reversing
changes made to 2d29a417803206a25c87764b6a7ca1d6cfab9afc.

Conflicts:
	lib/toolkit/stack-utils.js
