+++ lib/sdk/console/traceback.js
@@ -9,10 +9,10 @@ module.metadata = {
 };
 
 const { Cc, Ci, components } = require("chrome");
-const { parseStack, parseURI } = require("../../toolkit/stack-utils");
+const { parseStack, sourceURI } = require("toolkit/loader");
 const { readURISync } = require("../net/url");
 
-exports.sourceURI = parseURI;
+exports.sourceURI = sourceURI
 
 function safeGetFileLine(path, line) {
   try {
@@ -31,7 +31,7 @@ function nsIStackFramesToJSON(frame) {
   while (frame) {
     if (frame.filename) {
       stack.unshift({
-        fileName: parseURI(frame.filename),
+        fileName: sourceURI(frame.filename),
         lineNumber: frame.lineNumber,
         name: frame.name
       });
@@ -48,7 +48,7 @@ var fromException = exports.fromException = function fromException(e) {
   if (e.stack && e.stack.length)
     return parseStack(e.stack);
   if (e.fileName && typeof(e.lineNumber == "number"))
-    return [{fileName: parseURI(e.fileName),
+    return [{fileName: sourceURI(e.fileName),
              lineNumber: e.lineNumber,
              name: null}];
   return [];
