+++ lib/sdk/test/harness.js
@@ -9,7 +9,7 @@ module.metadata = {
 
 const { Cc, Ci, Cu } = require("chrome");
 const { Loader } = require('./loader');
-const { parseStack, serializeStack } = require("../../toolkit/stack-utils");
+const { serializeStack, parseStack  } = require("toolkit/loader");
 const { setTimeout } = require('../timers');
 const { PlainTextConsole } = require("../console/plain-text");
 const { when: unload } = require("../system/unload");
