+++ lib/toolkit/loader.js
@@ -42,8 +42,6 @@ const { NetUtil } = Cu.import("resource://gre/modules/NetUtil.jsm", {});
 const { Reflect } = Cu.import("resource://gre/modules/reflect.jsm", {});
 const { console } = Cu.import("resource://gre/modules/devtools/Console.jsm");
 const { join: pathJoin, normalize, dirname } = Cu.import("resource://gre/modules/osfile/ospath_unix.jsm");
-const toolkitURI = module.uri.replace(/loader\.js$/, "");
-const { parseStack, serializeStack, parseURI } = Cu.import(toolkitURI + "/stack-utils.js", {}).stackUtils;
 
 // Define some shortcuts.
 const bind = Function.call.bind(Function.bind);
@@ -130,12 +128,47 @@ const override = iced(function override(target, source) {
 });
 exports.override = override;
 
-// Export stack utils
-exports.parseStack = iced(parseStack);
-exports.serializeStack= iced(serializeStack);
-exports.sourceURI = iced(parseURI);
+function sourceURI(uri) { return String(uri).split(" -> ").pop(); }
+exports.sourceURI = iced(sourceURI);
 
 function isntLoaderFrame(frame) { return frame.fileName !== module.uri }
+
+function parseURI(uri) { return String(uri).split(" -> ").pop(); }
+exports.parseURI = parseURI;
+
+function parseStack(stack) {
+  let lines = String(stack).split("\n");
+  return lines.reduce(function(frames, line) {
+    if (line) {
+      let atIndex = line.indexOf("@");
+      let columnIndex = line.lastIndexOf(":");
+      let lineIndex = line.lastIndexOf(":", columnIndex - 1);
+      let fileName = parseURI(line.slice(atIndex + 1, lineIndex));
+      let lineNumber = parseInt(line.slice(lineIndex + 1, columnIndex));
+      let columnNumber = parseInt(line.slice(columnIndex + 1));
+      let name = line.slice(0, atIndex).split("(").shift();
+      frames.unshift({
+        fileName: fileName,
+        name: name,
+        lineNumber: lineNumber,
+        columnNumber: columnNumber
+      });
+    }
+    return frames;
+  }, []);
+}
+exports.parseStack = parseStack;
+
+function serializeStack(frames) {
+  return frames.reduce(function(stack, frame) {
+    return frame.name + "@" +
+           frame.fileName + ":" +
+           frame.lineNumber + ":" +
+           frame.columnNumber + "\n" +
+           stack;
+  }, "");
+}
+exports.serializeStack = serializeStack;
 
 function readURI(uri) {
   let stream = NetUtil.newChannel(uri, 'UTF-8', null).open();
@@ -278,7 +311,7 @@ const load = iced(function load(loader, module) {
     let stack = error.stack || Error().stack;
     let frames = parseStack(stack).filter(isntLoaderFrame);
     let toString = String(error);
-    let file = parseURI(fileName);
+    let file = sourceURI(fileName);
 
     // Note that `String(error)` where error is from subscript loader does
     // not puts `:` after `"Error"` unlike regular errors thrown by JS code.
@@ -590,7 +623,7 @@ const Require = iced(function Require(loader, requirer) {
     // at runtime.
     if (!(uri in modules)) {
       // Many of the loader's functionalities are dependent
-      // on modules[uri] being set before loading, so we set it and 
+      // on modules[uri] being set before loading, so we set it and
       // remove it if we have any errors.
       module = modules[uri] = Module(requirement, uri);
       try {
